backend/requirements.txt
requests
python-dotenv
fastapi
uvicorn
beautifulsoup4
pipecat-ai>=0.0.58
websockets>=15.0.0
langchain-openai
langchain-community
langchain
faiss-cpu
langchain_chroma
chromadb